Move adapters into their folder
