+++ src/adapters/ofxDepthCameraOrbbecAstra.h
